moved AbstractLocationObjective
