+++ src/main/java/org/betonquest/betonquest/compatibility/traincarts/objectives/TrainCartsLocationObjective.java
@@ -9,7 +9,7 @@ import org.betonquest.betonquest.instruction.Instruction;
 import org.betonquest.betonquest.instruction.argument.VariableArgument;
 import org.betonquest.betonquest.instruction.variable.VariableNumber;
 import org.betonquest.betonquest.instruction.variable.location.VariableLocation;
-import org.betonquest.betonquest.objective.AbstractLocationObjective;
+import org.betonquest.betonquest.quest.objective.location.AbstractLocationObjective;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.event.HandlerList;
