+++ src/main/java/org/betonquest/betonquest/compatibility/worldguard/RegionObjective.java
@@ -6,7 +6,7 @@ import org.betonquest.betonquest.api.profile.Profile;
 import org.betonquest.betonquest.api.quest.QuestException;
 import org.betonquest.betonquest.instruction.Instruction;
 import org.betonquest.betonquest.instruction.variable.VariableString;
-import org.betonquest.betonquest.objective.AbstractLocationObjective;
+import org.betonquest.betonquest.quest.objective.location.AbstractLocationObjective;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.event.HandlerList;
